% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_twostage_AdaptiveEnrichment_withrho.R
\name{decisionAdaptiveEnrichment}
\alias{decisionAdaptiveEnrichment}
\title{Performs hypothesis test for a clinical trial with adaptive enrichment}
\usage{
decisionAdaptiveEnrichment(lambda.2, w.2, x.1, x.2, r, sigma_1 = 1,
  sigma_2 = 1, rho = 0, lambda = 0.5, lambda.1 = 0.5, n = 100,
  w.1 = 0.5, alpha = 0.025, adj = c("bonferroni", "bonf",
  "bonferroni-holm", "simes"), gamma1 = ..gamma1..,
  gamma2 = ..gamma2.., gamma12 = ..gamma12.., fixed.weights = FALSE,
  verbose = TRUE, show.case = TRUE)
}
\arguments{
\item{x.1}{1st stage data}

\item{x.2}{2nd stage data}

\item{lambda}{proportion of patients in sub-population 1 in the overal patient population}

\item{lambda.1}{proportion of patients in sub-population 1 in the trial 1st stage}

\item{n}{trial sample size}

\item{w.1}{factor for allocation of alpha to H01 in the testing correction (1st stage)}

\item{alpha}{significance level}

\item{gamma1}{gain when rejecting H01. A function with parameters (mu1, mu2, theta1, theta2, lambda)}

\item{gamma2}{gain when rejecting H02. A function with parameters (mu1, mu2, theta1, theta2, lambda)}

\item{gamma12}{gain when rejecting H01 and H02. A function with parameters (mu1, mu2, theta1, theta2, lambda)}

\item{verbose}{Print intermediate values for calculations}

\item{r.1}{proportion of patients in 1st stage of the trial}

\item{theta_1}{true parameter value for sub-population 1}

\item{theta_2}{true parameter value for sub-population 2}

\item{mu_1}{prior mean for the treatment effect in sub-population 1}

\item{psi_1}{prior standard deviation for the treatment effect in sub-population 1}

\item{mu_2}{prior mean for the treatment effect in sub-population 2}

\item{psi_2}{prior standard deviation for the treatment effect in sub-population 2}

\item{rho_0}{prior correlation between treatment effect in sub-population 1 and sub-population 2 (ignored)}

\item{tol.zero}{Absolute distance to 0 or 1 to be considered 0 or 1}

\item{use.prior.as.true}{logical. If TRUE, then the true theta is generate using the prior distribution and theta_1 and theta_2 are not used.}
}
\description{
Given values for first and second stage estimates, the function generates
test the hypotheses of no effect in the subgroup
}
