% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_twostage_withrho.R
\name{decisionOptimalTrial}
\alias{decisionOptimalTrial}
\title{Performs the statistical test using first and second stage data}
\usage{
decisionOptimalTrial(lambda.2, w.2, x.1, x.2, r, sigma_1 = 1,
  sigma_2 = 1, rho = 0, lambda = 0.5, lambda.1 = 0.5, n = 100,
  w.1 = 0.5, alpha = 0.025, gamma1 = ..gamma1..,
  gamma2 = ..gamma2.., gamma12 = ..gamma12.., fixed.weights = FALSE,
  verbose = TRUE, show.case = TRUE)
}
\arguments{
\item{lambda.2}{proportion of patients in sub-population 1 in the trial 2nd stage}

\item{w.2}{factor for allocation of alpha to H01 in the testing correction (final testing)}

\item{x.1}{the values of the first stage estimates thetahat.1_1 and thetahat.1_2}

\item{x.2}{the values of the second stage estimates  thetahat.2_1 and thetahat.2_2}

\item{r}{proportion of patients in 1st stage of the trial}

\item{sigma_1}{standard deviation in subpopulation 1}

\item{sigma_2}{standard deviation in subpopulation 2}

\item{rho}{correlation between estimates in subpopulation 1 and subpopulation 2 (ignored)}

\item{lambda}{proportion of patients in sub-population 1 in the overal patient population}

\item{lambda.1}{proportion of patients in sub-population 1 in the trial 1st stage}

\item{n}{trial sample size}

\item{w.1}{factor for allocation of alpha to H01 in the testing correction (1st stage)}

\item{alpha}{significance level}

\item{gamma1}{gain when rejecting H01. A function with parameters (mu1, mu2, theta1, theta2, lambda)}

\item{gamma2}{gain when rejecting H02. A function with parameters (mu1, mu2, theta1, theta2, lambda)}

\item{gamma12}{gain when rejecting H01 and H02. A function with parameters (mu1, mu2, theta1, theta2, lambda)}

\item{verbose}{Print intermediate values for calculations}

\item{tol.zero}{Absolute distance to 0 or 1 to be considered 0 or 1}
}
\description{
See manuscript
}
