% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_twostage_withrho.R
\name{interimOptimalTrial_EU}
\alias{interimOptimalTrial_EU}
\title{Expected utility for values of lambda.2 and w.2 after 1st stage}
\usage{
interimOptimalTrial_EU(lambda.2, w.2, x, r, sigma_1 = 1, sigma_2 = 1,
  rho = 0, mu_1 = 0, psi_1 = 1, mu_2 = 0, psi_2 = 1, rho0 = 0,
  lambda = 0.5, lambda.1 = 0.5, n = 100, w.1 = 0.5,
  alpha = 0.025, gamma1 = ..gamma1.., gamma2 = ..gamma2..,
  gamma12 = ..gamma12.., verbose = TRUE, show.case = TRUE,
  tol.zero = 1e-10)
}
\arguments{
\item{lambda.2}{proportion of patients in sub-population 1 in the trial 2nd stage}

\item{x}{the values of the first stage estimates theta1 and theta2}

\item{r}{proportion of patients in 1st stage of the trial}

\item{sigma_1}{standard deviation in subpopulation 1}

\item{sigma_2}{standard deviation in subpopulation 2}

\item{rho}{correlation between estimates in subpopulation 1 and subpopulation 2 (ignored)}

\item{mu_1}{prior mean for the treatment effect in sub-population 1}

\item{psi_1}{prior standard deviation for the treatment effect in sub-population 1}

\item{mu_2}{prior mean for the treatment effect in sub-population 2}

\item{psi_2}{prior standard deviation for the treatment effect in sub-population 2}

\item{lambda}{proportion of patients in sub-population 1 in the overal patient population}

\item{lambda.1}{proportion of patients in sub-population 1 in the trial 1st stage}

\item{n}{trial sample size}

\item{w.1}{factor for allocation of alpha to H01 in the testing correction}

\item{alpha}{significance level}

\item{gamma1}{gain when rejecting H01. a function with parameters (mu1, mu2, theta1, theta2, lambda)}

\item{gamma2}{gain when rejecting H02. a function with parameters (mu1, mu2, theta1, theta2, lambda)}

\item{gamma12}{gain when rejecting H01 and H02. a function with parameters (mu1, mu2, theta1, theta2, lambda)}

\item{verbose}{Print intermediate values for calculations}

\item{tol.zero}{Absolute distance to 0 or 1 to be considered 0 or 1}

\item{rho_0}{prior correlation between treatment effect in sub-population 1 and sub-population 2 (ignored)}
}
\description{
Simplified version where the utility function does not depend on the estimates or the true
values of the parameters and can be taken out of the first integration
}
