% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_umbrellaTrial-twoStage.R
\name{interimUmbrellaTrial_grid}
\alias{interimUmbrellaTrial_grid}
\title{Find the optimized lambda.2 for the trial}
\usage{
interimUmbrellaTrial_grid(x = NULL, r = 0.5, sigma_1 = 1,
  sigma_2 = 1, sigma_3 = 1, rho = 0, mu_1 = 0, psi_1 = 1,
  mu_2 = 0, psi_2 = 1, mu_3 = 0, psi_3 = 1, rho0 = 0,
  lambda = c(0.15, 0.05), lambda.1 = c(0.15, 0.05), n = 1484,
  alpha = c(0.05, 0.05, 0.025), gamma1 = ..umbrella_gamma1..,
  gamma2 = ..umbrella_gamma2.., gamma3 = ..umbrella_gamma3..,
  verbose = FALSE, tol.zero = 1e-10, tol.lim = 1e-04,
  optim_func = "optim", sig.digits = 4)
}
\arguments{
\item{x}{the values of the first stage estimates theta1 and theta2}

\item{r}{proportion of patients in 1st stage of the trial}

\item{sigma_1}{standard deviation in subpopulation 1}

\item{sigma_2}{standard deviation in subpopulation 2}

\item{rho}{correlation between estimates in subpopulation 1 and subpopulation 2 (ignored)}

\item{mu_1}{prior mean for the treatment effect in sub-population 1}

\item{psi_1}{prior standard deviation for the treatment effect in sub-population 1}

\item{mu_2}{prior mean for the treatment effect in sub-population 2}

\item{psi_2}{prior standard deviation for the treatment effect in sub-population 2}

\item{lambda}{proportion of patients in sub-population 1 in the overal patient population}

\item{lambda.1}{proportion of patients in sub-population 1 in the trial 1st stage}

\item{n}{trial sample size}

\item{alpha}{significance level}

\item{gamma1}{gain when rejecting H01. A function with parameters (mu1, mu2, theta1, theta2, lambda)}

\item{gamma2}{gain when rejecting H02. A function with parameters (mu1, mu2, theta1, theta2, lambda)}

\item{verbose}{Print intermediate values for calculations}

\item{tol.zero}{Absolute distance to 0 or 1 to be considered 0 or 1}

\item{rho_0}{prior correlation between treatment effect in sub-population 1 and sub-population 2 (ignored)}

\item{w.1}{factor for allocation of alpha to H01 in the testing correction}

\item{gamma12}{gain when rejecting H01 and H02. A function with parameters (mu1, mu2, theta1, theta2, lambda)}
}
\description{
Uses the interimUmbrellaTrial_EU function to find the values that yield to the maximum utility.
Simplified version when the utility functino does not depend on the estimates or the true
values of the parameters and can be taken out of the first integration
}
