% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_twostage_withrho.R
\name{optimalMtrials}
\alias{optimalMtrials}
\title{Simulates M clinical trials with optimized sample allocation using simulate1trial}
\usage{
optimalMtrials(nsim = 1000, mc.cores = 1, r.1 = 0.1, lambda = 0.5,
  lambda.1 = 0.5, w.1 = 0.5, theta_1 = 0, theta_2 = 0, mu_1 = 0,
  psi_1 = 1, mu_2 = 0, psi_2 = 1, rho0 = 0, alpha = 0.05,
  n = 100, gamma1 = ..gamma1.., gamma2 = ..gamma2..,
  gamma12 = ..gamma12.., tol.zero = 1e-05, tol.lim = 1e-04,
  verbose = TRUE, summarize = FALSE, use.prior.as.true = FALSE,
  optim_func = "optim")
}
\description{
Simulates M clinical trials with optimized sample allocation using simulate1trial
}
