% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_umbrella2armTrial-twoStage.R
\name{umbrella2arm1Trial}
\alias{umbrella2arm1Trial}
\title{Simulates a clinical trial with optimized sample allocation.}
\usage{
umbrella2arm1Trial(r = 0.5, theta_1 = 0, theta_2 = 0, sigma_1 = 1,
  sigma_2 = 1, rho = 0, mu_1 = 0, psi_1 = 1, mu_2 = 0,
  psi_2 = 1, rho0 = 0, lambda = c(0.15), lambda.1 = c(0.15),
  n = 1484, alpha = c(0.05, 0.05), gamma1 = ..umbrella_gamma1..,
  gamma2 = ..umbrella_gamma2.., verbose = TRUE, tol.zero = 1e-10,
  use.prior.as.true = FALSE, sig.digits = 4)
}
\arguments{
\item{theta_1}{true parameter value for sub-population 1}

\item{theta_2}{true parameter value for sub-population 2}

\item{mu_1}{prior mean for the treatment effect in sub-population 1}

\item{psi_1}{prior standard deviation for the treatment effect in sub-population 1}

\item{mu_2}{prior mean for the treatment effect in sub-population 2}

\item{psi_2}{prior standard deviation for the treatment effect in sub-population 2}

\item{lambda}{proportion of patients in sub-population 1 in the overal patient population}

\item{lambda.1}{proportion of patients in sub-population 1 in the trial 1st stage}

\item{n}{trial sample size}

\item{alpha}{significance level}

\item{gamma1}{gain when rejecting H01. A function with parameters (mu1, mu2, theta1, theta2, lambda)}

\item{gamma2}{gain when rejecting H02. A function with parameters (mu1, mu2, theta1, theta2, lambda)}

\item{verbose}{Print intermediate values for calculations}

\item{tol.zero}{Absolute distance to 0 or 1 to be considered 0 or 1}

\item{use.prior.as.true}{logical. If TRUE, then the true theta is generate using the prior distribution and theta_1 and theta_2 are not used.}

\item{r.1}{proportion of patients in 1st stage of the trial}

\item{w.1}{factor for allocation of alpha to H01 in the testing correction (1st stage)}

\item{rho_0}{prior correlation between treatment effect in sub-population 1 and sub-population 2 (ignored)}

\item{gamma12}{gain when rejecting H01 and H02. A function with parameters (mu1, mu2, theta1, theta2, lambda)}
}
\description{
Given true values for theta_1 and theta_2, the function generates first stage data,
then find optimal values lambda.2 and w.2, generates second stage data, and
finally test the hypotheses.
}
