% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_umbrella2armTrial-twoStage.R
\name{umbrella2armMTrials}
\alias{umbrella2armMTrials}
\title{Simulates M clinical trials with optimized sample allocation using simulate1trial}
\usage{
umbrella2armMTrials(nsim = 1000, mc.cores = 1, r = 0.5,
  theta_1 = 0, theta_2 = 0, sigma_1 = 1, sigma_2 = 1, rho = 0,
  mu_1 = 0, psi_1 = 1, mu_2 = 0, psi_2 = 1, rho0 = 0,
  lambda = c(0.15), lambda.1 = c(0.15), n = 1484, alpha = c(0.05,
  0.05), gamma1 = ..umbrella_gamma1.., gamma2 = ..umbrella_gamma2..,
  verbose = FALSE, tol.zero = 1e-10, summarize = TRUE,
  use.prior.as.true = FALSE, sig.digits = 4)
}
\description{
Simulates M clinical trials with optimized sample allocation using simulate1trial
}
