% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_umbrellaTrial-twoStage.R
\name{umbrellaFullOptimization}
\alias{umbrellaFullOptimization}
\title{Helps find optimal values of first stage prevalences}
\usage{
umbrellaFullOptimization(M = 1000, mc.cores = 40, r = 0.5,
  sigma_1 = 1, sigma_2 = 1, sigma_3 = 1, rho = 0, mu_1 = 0,
  psi_1 = 1, mu_2 = 0, psi_2 = 1, mu_3 = 0, psi_3 = 1,
  rho0 = 0, lambda = c(0.15, 0.05), lambda.1 = c(0.15, 0.05),
  n = 1484, alpha = c(0.05, 0.05, 0.025),
  gamma1 = ..umbrella_gamma1.., gamma2 = ..umbrella_gamma2..,
  gamma3 = ..umbrella_gamma3.., verbose = TRUE, tol.zero = 1e-10,
  sig.digits = 4, summarize = TRUE)
}
\arguments{
\item{mu_1}{prior mean for the treatment effect in sub-population 1}

\item{psi_1}{prior standard deviation for the treatment effect in sub-population 1}

\item{mu_2}{prior mean for the treatment effect in sub-population 2}

\item{psi_2}{prior standard deviation for the treatment effect in sub-population 2}

\item{lambda}{proportion of patients in sub-population 1 in the overal patient population}

\item{lambda.1}{proportion of patients in sub-population 1 in the trial 1st stage}

\item{n}{trial sample size}

\item{alpha}{significance level}

\item{gamma1}{gain when rejecting H01. A function with parameters (mu1, mu2, theta1, theta2, lambda)}

\item{gamma2}{gain when rejecting H02. A function with parameters (mu1, mu2, theta1, theta2, lambda)}

\item{verbose}{Print intermediate values for calculations}

\item{tol.zero}{Absolute distance to 0 or 1 to be considered 0 or 1}

\item{r.1}{proportion of patients in 1st stage of the trial}

\item{w.1}{factor for allocation of alpha to H01 in the testing correction (1st stage)}

\item{theta_1}{true parameter value for sub-population 1}

\item{theta_2}{true parameter value for sub-population 2}

\item{rho_0}{prior correlation between treatment effect in sub-population 1 and sub-population 2 (ignored)}

\item{gamma12}{gain when rejecting H01 and H02. A function with parameters (mu1, mu2, theta1, theta2, lambda)}
}
\description{
We calculate expected utility of the trial given that we will optimize at interim.
}
