% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_umbrellaTrial-twoStage.R
\name{umbrellaMTrials}
\alias{umbrellaMTrials}
\title{Simulates M clinical trials with optimized sample allocation using simulate1trial}
\usage{
umbrellaMTrials(nsim = 1000, mc.cores = 1, r = 0.5, theta_1 = 0,
  theta_2 = 0, theta_3 = 0, sigma_1 = 1, sigma_2 = 1,
  sigma_3 = 1, rho = 0, mu_1 = 0, psi_1 = 1, mu_2 = 0,
  psi_2 = 1, mu_3 = 0, psi_3 = 1, rho0 = 0, lambda = c(0.15,
  0.05), lambda.1 = c(0.15, 0.05), n = 1484, alpha = c(0.05, 0.05,
  0.025), gamma1 = ..umbrella_gamma1.., gamma2 = ..umbrella_gamma2..,
  gamma3 = ..umbrella_gamma3.., verbose = FALSE, tol.zero = 1e-10,
  summarize = TRUE, use.prior.as.true = FALSE, sig.digits = 4)
}
\description{
Simulates M clinical trials with optimized sample allocation using simulate1trial
}
